requirements-dev.txt
# Development Dependencies
# Install with: pip install -r requirements-dev.txt

# Base requirements
-r backend/requirements.txt

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.2.0

# Code Quality
black==23.7.0
flake8==6.0.0
pylint==2.17.5
isort==5.12.0

# Debugging
python-debugpy==1.6.7
ipython==8.14.0

# Development Server
python-dotenv==1.0.0
Werkzeug==2.3.7

# Documentation
Sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Type Checking
mypy==1.4.1
types-requests==2.31.0.2

# API Documentation
flask-restx==0.5.1

# Database Utilities
flask-shell-ipython==0.4.0

# Profiling
flask-debugtoolbar==0.13.1
py-spy==0.3.14
